feat: article-meta - add changes from PR
