+++ packages/article-meta/article-meta.stories.js
@@ -1,5 +1,4 @@
 import React from "react";
-import { View } from "react-native";
 import { storiesOf } from "@storybook/react-native";
 import ArticleMeta from "./article-meta";
 
@@ -8,8 +7,6 @@ const props = {
   publication: "The Sunday Times"
 };
 
-const story = m => <View style={{ padding: 20 }}>{m}</View>;
-
 storiesOf("ArticleMeta", module).add("ArticleMeta", () =>
-  story(<ArticleMeta {...props} />)
+  <ArticleMeta {...props} />
 );
